Split tool integration tests
