+++ .circleci/config.yml
@@ -74,6 +74,7 @@ workflows:
               only:
                 - main
                 - staging
+                - devops/pytest-parallelism
   deploy:
     jobs:
       - deploy-to-pypi:
